test: moving test-cli.sh to golang test (#480)
## Context

Some end to end tests are executed with bash scripts - others are executed with golang. 

This mix and match makes it harder to maintain and to think about. 

## Changes

This PR moves the `test-cli.sh` test script to an equivalent golang script. 

The golang script compares each command separately, instead of concataining all results to a single golden file.